refactor: useBaseField 분리 (useInputValue 훅으로 변경 / 상태 값만 관리)
